Move JedisMetaInfo into parent package and make it package private
